move getargspec, sentinel into traitlets/utils
