+++ traitlets/utils/sentinel.py
